Update pedido e pedido detalhe

+++ src/main/java/com/learning/api/minisys/dtos/faturamento/pedido/PedidoDetalheDto.java
@@ -11,7 +11,7 @@ public record PedidoDetalheDto(
         NewPedidoDto pedido,
         NewItemDto item,
         Boolean cancelado,
-        int ordem,
+        Long ordem,
         @NotBlank(message = "O campo descrição é obrigatório")
         String descricao,
         NewUnidadeMedidaDto unidadeMedida,
